Move RPC server task queue to separate file
